Adds permissions support to RedstoneChips.
Notes:
New preference added: "usePermissions". If set to true, permissions will be used. When set to false, everything behaves as it does today except for "protected channels" detailed below.

All commands have an "allow" and "deny" permission. The "deny" permission overrides the "allow" permission. This can be used if a user belongs to a group which allows a command that you need to explicitly deny the user. Additionally, RCreset supports an "all" permission allowing "rcreset all" and RCprefs supports a "set" permission allowing the changing of preferences, both also support a "deny" permission. RCp does not support permissions.

All circuits have create and destroy permissions. These also have an "allow" and "deny", which function identically as above. The author of new circuits does not need to add anything, all circuit types are automatically supported. Destruction protection will prevent a circuit from being destroyed via RC commands or by breaking the blocks that make up the circuit.

Also, support for "protected channels" has been added. This allows users to control who can create and destroy wireless chips on particular channels. This system is independent of the other permissions and functions even when permissions are disabled. Users can define admins for the channel, who can change who has access to the channel, and users, who can create chips using the channel, but can not change access. RCchannels will not show a user any channels that they do not have access to, and protected channels the user does have access to will be shown with a "P" to identify they are protected. RCchannels <Channel Name> will show all authorized users and admins of the channel. Additionally, a preference is provided which allows a user to act as an admin for all channels. A new command "RCprotect" has been added to manage these channels. Some additional data is saved to the redstonechips.circuits file to hold the channel information.

All pre-defined permissions and documentation for the new command can be found in the plugin.yml file. Also several permission groups are defined in the plugin.yml to make configuration of permissions easier. Individual circuit permissions are dynamic and not included in the plugin.yml. These follow the format of "redstonechips.circuit.create.<CIRCUIT NAME>", "redstonechips.circuit.destroy.<CIRCUIT NAME>" for the allow permissions and "redstonechips.circuit.create.<CIRCUIT NAME>.deny", "redstonechips.circuit.destroy.<CIRCUIT NAME>.deny" for the "deny" permissions.

All permissions are checked for any given operation. For example if resetting a transmitter via the "rcreset" command:
1. Permission for rcreset is checked.
2. Permission for breaking (destroying) transmitters is checked.
3. Permission for destroying transmitters on that channel is checked.
4. Permission for activating transmitters is checked.
5. Permission for activating transmitters on that channel is checked.

This will only work with CB 1000 or higher.

HMMM... I think that is it.